Improve internal threadsafety (#25)
